refactor: factored out ParameterReadingRange
